notebooks/style-swin-gan/requirements.txt
torch
torchvision
torchmetrics
pillow
cudatoolkit-dev
